# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LibTrack System API",
        "description": "API documentation for the LibTrack Library Tracking System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/books": {
            "get": {
                "tags": [
                    "Book Management"
                ],
                "summary": "Get all books",
                "operationId": "3928b0b6916b8d707bd9de0c7f7225be",
                "responses": {
                    "200": {
                        "description": "Books retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Book Management"
                ],
                "summary": "Add a new book",
                "operationId": "d6741a88ba80015e1bfefdc77c8c813e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "isbn",
                                    "category",
                                    "quantity"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Introduction to Networking"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Jain Alawi"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "978-1234567890"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Networking"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A basic introduction to computer networking."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book added successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/books/{id}": {
            "get": {
                "tags": [
                    "Book Management"
                ],
                "summary": "Get a single book",
                "operationId": "e9679e2bc7bda8c467a7ff22d24e9187",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book retrieved successfully"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Book Management"
                ],
                "summary": "Update a book",
                "operationId": "44f7915e0c88331f8aebe4478bc0f6d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "isbn": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated successfully"
                    },
                    "404": {
                        "description": "Book not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Book Management"
                ],
                "summary": "Delete a book",
                "operationId": "6f5009124f6d6459be39acfb5555f91f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book deleted successfully"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer Token",
                "scheme": "bearer"
            }
        }
    },
...
}